hexo_circle_of_friends/requirement.txt
arrow==1.2.1
attrs==21.2.0
Automat==20.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
constantly==15.1.0
cryptography==36.0.0
cssselect==1.1.0
gevent==21.8.0
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
iso8601==1.0.2
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
leancloud==2.9.7
lxml==4.6.4
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
python-dateutil==2.8.2
qiniu==7.3.1
queuelib==1.6.2
requests==2.26.0
Scrapy==2.5.1
secure-cookie==0.2.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
Twisted==21.7.0
typing-extensions==4.0.0
urllib3==1.26.7
w3lib==1.22.0
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.4.0
PyYAML==5.3.1
feedparser==6.0.8